+++ branches/GZDoomBuilder/Source/Plugins/BuilderModes/VisualModes/BaseVisualGeometrySidedef.cs
@@ -158,7 +158,7 @@ namespace CodeImp.DoomBuilder.BuilderModes
 		}
 		
 		// Processing
-		public virtual void OnProcess(double deltatime)
+		public virtual void OnProcess(float deltatime)
 		{
 			// If the texture was not loaded, but is loaded now, then re-setup geometry
 			if(setuponloadedtexture != 0)
